Updates Issue#1
Rather than adding first/last tables to calculate the time
elapsed for loading, we use the creation time of the first table (employees)
and the update time of the last table (salaries) to calculate the time.

A similar operation was added to the test_* files, so that they can
show how long the checksum operation takes.